ingreso de funcionalidades en las distintas clases

+++ src/main/java/com/mycompany/herenciaypolimorfismo/Empleado.java
@@ -104,7 +104,7 @@ public class Empleado {
     
     
     //aumenta slario en base a cantidad de dinero
-    public void incrementarSalario(int num){
+    public void incrementarSalario(long num){
         this.salario = this.salario + num;
     }
